+++ superdesk/text_checkers/spellcheckers/__init__.py
@@ -22,6 +22,12 @@ logger = logging.getLogger(__name__)
 # can be set to False if importSpellcheckers need to be called manually
 # (e.g. in unit tests)
 AUTO_IMPORT = True
+# default spellchecker is the basic one, using internal dictionary
+SPELLCHECKER_DEFAULT = "default"
+CAP_SPELLING = "spelling"
+CAP_GRAMMAR = "grammar"
+# any language is supported
+LANG_ANY = "*"
 
 
 class SpellcheckerResource(Resource):
@@ -34,9 +40,18 @@ class SpellcheckerResource(Resource):
             'type': 'string',
             'required': True,
         },
+        'language': {
+            'type': 'string',
+            'nullable': True,
+        },
         'suggestions': {
             'type': 'boolean',
             'default': False,
+        },
+        # if True, spelling errors which are in internal dictionary will be ignored
+        'use_internal_dict': {
+            'type': 'boolean',
+            'default': True,
         },
     }
     internal_resource = False
@@ -45,22 +60,85 @@ class SpellcheckerResource(Resource):
 
 
 class SpellcheckerService(BaseService):
+    r"""Service managing spellchecking and suggestions.
+
+    When doing a POST request on this service, the following keys can be used (keys with a \* are required):
+
+
+    =================   ===========
+    key                 explanation
+    =================   ===========
+    spellchecker \*     name of the spellchecker
+    text \*             text to check or word for suggestions
+    language            language to use (for spellcheckers handling several ones)
+    suggestions         false (default) to check a text, else will get suggestions
+    use_internal_dict   true (default) to remove spellings mistakes from words in personal dictionary
+    =================   ===========
+
+    e.g. to check a French text while removing words from personal dictionary (grammar mistakes are here on purpose):
+
+    .. sourcecode:: json
+
+        {
+            "spellchecker": "grammalecte",
+            "text": "Il nous reste à vérifié votre maquette."
+        }
+
+    Here is an example payload for suggestions:
+
+    .. sourcecode:: json
+
+        {
+            "spellchecker": "grammalecte",
+            "suggestions": true,
+            "text": "fote"
+        }
+
+    """
+
+    def remove_errors_in_dict(self, spellchecker, language, check_data):
+        """Remove spelling error which are in the internal dictionary
+
+        :param SpellcheckerBase: spellchecker used
+        :param str language: language specified by the client
+        :param dict check_data: data returned by spellchecker.check method (will be modified in place)
+        """
+        if spellchecker.name == SPELLCHECKER_DEFAULT:
+            # default spellchecker already works with internal dictionaries
+            return
+        errors = check_data['errors']
+        if not errors:
+            return
+        lang = spellchecker.get_language(language)
+        dictionaries_service = superdesk.get_resource_service('dictionaries')
+        model = dictionaries_service.get_model_for_lang(lang)
+        to_remove = []
+        for error in errors:
+            if error.get('type', 'spelling') != 'spelling':
+                continue
+            if error['text'].lower() in model:
+                to_remove.append(error)
+        for error in to_remove:
+            errors.remove(error)
 
     def create(self, docs, **kwargs):
         # we override create because we don't want anything stored in database
         doc = docs[0]
         sc_name = doc["spellchecker"]
+        language = doc.get('language')
         try:
             spellchecker = registered_spellcheckers[sc_name]
         except KeyError:
             raise SuperdeskApiError.notFoundError("{sc_name} spellchecker can't be found".format(sc_name=sc_name))
 
         if doc["suggestions"]:
-            check_data = spellchecker.suggest(doc['text'])
+            check_data = spellchecker.suggest(doc['text'], language)
             assert "suggestions" in check_data
         else:
-            check_data = spellchecker.check(doc['text'])
+            check_data = spellchecker.check(doc['text'], language)
             assert "errors" in check_data
+            if doc["use_internal_dict"]:
+                self.remove_errors_in_dict(spellchecker, language, check_data)
         docs[0].update(check_data)
         return [0]